requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
asyncio==3.4.3
attrs==23.2.0
chardet==4.0.0
discord.py==2.3.2
idna==3.6
multidict==6.0.5
typing_extensions==4.10.0
yarl==1.9.4
python-jobspy==1.1.51
python-dotenv==1.0.1
tweepy==4.14.0
PyYAML==6.0.1
